+++ client/src/pages/ems-fd/dash.tsx
@@ -29,9 +29,13 @@ const EmsFdDash: React.FC<Props> = (props) => {
   useDocTitle("EMS-FD Dashboard");
 
   React.useEffect(() => {
-    socket.on("UPDATE_AOP", (newAop: string) => {
-      setAop(newAop);
-    });
+    const handler = (newAop: string) => setAop(newAop);
+
+    socket.on("UPDATE_AOP", handler);
+
+    return () => {
+      socket.off("UPDATE_AOP", handler);
+    };
   }, []);
 
   React.useEffect(() => {